requirements.txt
absl-py==2.1.0
altair==5.5.0
altgraph==0.17.4
#array_record==0.5.1 not available
array_record
astunparse==1.6.3
attrs==24.2.0
audioread==3.0.1
backports.tarfile==1.2.0
black==24.10.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
docstring_parser==0.16
docutils==0.21.2
etils==1.10.0
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
h5py==3.12.1
idna==3.10
immutabledict==4.2.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kagglehub==0.3.4
keras==2.15.0
keras-core==0.1.7
keras-cv==0.9.0
keyring==25.5.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==10.5.0
msgpack==1.1.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.14.2
nh3==0.2.18
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
pkginfo==1.10.0
platformdirs==4.3.6
plotly==5.24.1
pooch==1.8.2
pre==1.3.0
promise==2.3
protobuf==4.25.5
psutil==6.1.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.0
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
readme_renderer==44.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
SecretStorage==3.3.3
simple-parsing==0.1.6
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.5.0.post1
streamlit==1.40.2
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.7
tensorflow-estimator==2.15.0
tensorflow-intel==0.0.1
tensorflow-io==0.37.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
twine==5.1.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.14.1
zipp==3.21.0
PySocks==1.7.1 
beautifulsoup4==4.13.3 
filelock==3.17.0 
gdown==5.2.0 
soupsieve==2.6